+++ tests/helpers/request.js
@@ -29,10 +29,25 @@ class Request {
 	 * @see [[BrowserTests.Request.waitForRequests]]
 	 */
 	waitForRequests(page, urls) {
+		return this.waitForRequestsEvents(page, urls, 'request');
+	}
+
+	/**
+	 * @see [[BrowserTests.Request.waitForRequestsFail]]
+	 */
+	waitForRequestsFail(page, urls) {
+		return this.waitForRequestsEvents(page, urls, 'requestfailed');
+	}
+
+	/**
+	 * @see [[BrowserTests.Request.waitForRequestsEvents]]
+	 * @private
+	 */
+	waitForRequestsEvents(page, urls, event) {
 		const normalized = [...urls];
 
 		return new Promise((res) => {
-			page.on('request', (request) => {
+			page.on(event, (request) => {
 				const index = normalized.indexOf(request.url());
 
 				if (index === -1) {
@@ -47,6 +62,42 @@ class Request {
 			});
 
 		});
+	}
+
+	/**
+	 * @see [[BrowserTests.Request.interceptRequests]]
+	 */
+	interceptRequests(page, urls, response, timeout) {
+		return Promise.all(urls.map((url) => this.interceptRequest(page, url, response, timeout)));
+	}
+
+	/**
+	 * @see [[BrowserTests.Request.interceptRequest]]
+	 */
+	interceptRequest(page, url, response, timeout) {
+		return new Promise((res, rej) => {
+			const handler = (route) => {
+				route.fulfill({status: 200, ...response});
+				page.unroute(url, handler);
+				res();
+			};
+
+			if (timeout != null) {
+				setTimeout(() => {
+					page.unroute(url, handler);
+					rej();
+				}, timeout);
+			}
+
+			page.route(url, handler);
+		});
+	}
+
+	/**
+	 * @see [[BrowserTests.Request.getRandomUrl]]
+	 */
+	getRandomUrl() {
+		return `https://v4fire-random-url.com/${String(Math.random()).substring(4)}`;
 	}
 
 	/**