+++ src/super/i-data/CHANGELOG.md
@@ -9,6 +9,29 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+##   v3.0.0-rc.72 ()
+
+#### :boom: Breaking Change
+
+* Renamed `dataProviderEmitter` to `dataEmitter`
+* Deprecated `requestFilter`
+* Deprecated `dropRequestCache`
+
+#### :rocket: New Feature
+
+* Added `defaultRequestFilter`
+* Added `dropDataCache`
+
+#### :nail_care: Polish
+
+* Improved doc
+
+## v3.0.0-rc.61 (2020-09-04)
+
+#### :bug: Bug Fix
+
+* [Fixed wrong refactoring in `rc48`](https://github.com/V4Fire/Client/pull/326)
+
 ## v3.0.0-rc.56 (2020-08-06)
 
 #### :bug: Bug Fix